[BEAM-3213] Refactor: remove test code duplication
